Update roboclaw_private.go
Restored the buffer for reading responses from the after a write to confirm write success. I replaced this months ago with a timeout instead to prevent any delays if the RX response failed to be received. However, since making this library public I have switched it back to waiting for a response, similar to the arduino library. To mitigate the above risk the maxretry should be configurable.